MVP 1.0 complete, bugfixes, and scheduler implementation.

+++ src/main/java/nz/co/twg/eim/model/condition/Condition.java
@@ -10,7 +10,7 @@ public interface Condition<T> extends EimObject {
 
     String getDirectory();
 
-    Optional<Integer> getMaxAge();
+    Optional<Long> getMaxAge();
 
     Optional<Integer> getFileCount();
 }
